requirements.txt
pysam
synbiochem-py